+++ migrations/db/init-scripts/00000000000000-initial-schema.sql
@@ -1,23 +1,23 @@
 -- migrate:up
-\set pgpass `echo "$POSTGRES_PASSWORD"`
+\set db `echo "$POSTGRES_DB"`
 
 -- Set up realtime
 -- defaults to empty publication
 create publication supabase_realtime;
 
 -- Supabase super admin
-alter user  supabase_admin with superuser createdb createrole replication bypassrls;
+-- alter user  supabase_admin with superuser createdb createrole replication bypassrls;
 
 -- Supabase replication user
-create user supabase_replication_admin with login replication password :'pgpass';
+-- create user supabase_replication_admin with login replication;
 
 -- Supabase etl user
-create user supabase_etl_admin with login replication;
-grant pg_read_all_data to supabase_etl_admin;
-grant create on database postgres to supabase_etl_admin;
+-- create user supabase_etl_admin with login replication;
+-- grant pg_read_all_data to supabase_etl_admin;
+grant create on database :db to supabase_etl_admin;
 
 -- Supabase read-only user
-create role supabase_read_only_user with login bypassrls password :'pgpass';
+-- create role supabase_read_only_user with login bypassrls;
 -- grant pg_read_all_data to supabase_read_only_user;
 
 -- Extension namespacing
@@ -29,11 +29,11 @@ create extension if not exists postgis          with schema extensions;
 
 
 -- Set up auth roles for the developer
-create role anon                nologin inherit;
-create role authenticated       nologin inherit; -- "logged in" user: web_user, app_user, etc
-create role service_role        nologin inherit bypassrls; -- allow developers to create JWT's that bypass their policies
+-- create role anon                nologin inherit;
+-- create role authenticated       nologin inherit; -- "logged in" user: web_user, app_user, etc
+-- create role service_role        nologin inherit bypassrls; -- allow developers to create JWT's that bypass their policies
 
-create user authenticator noinherit password :'pgpass';
+-- create user authenticator noinherit;
 grant anon              to authenticator;
 grant authenticated     to authenticator;
 grant service_role      to authenticator;